Replace `GlobalContext` by `HeaderContext`
