+++ src/Column/ActionColumnRenderer.php
@@ -10,7 +10,6 @@ use Yiisoft\Html\Html;
 use Yiisoft\Yii\DataView\Column\Base\Cell;
 use Yiisoft\Yii\DataView\Column\Base\GlobalContext;
 use Yiisoft\Yii\DataView\Column\Base\DataContext;
-use Yiisoft\Yii\DataView\Column\Base\HeaderContext;
 
 use function is_bool;
 use function is_callable;
@@ -73,7 +72,7 @@ final class ActionColumnRenderer implements ColumnRendererInterface
         return $cell->addAttributes($column->columnAttributes);
     }
 
-    public function renderHeader(ColumnInterface $column, Cell $cell, HeaderContext $context): Cell
+    public function renderHeader(ColumnInterface $column, Cell $cell, GlobalContext $context): Cell
     {
         return $cell
             ->content($column->header ?? $context->translate('Actions'))
